refactor: add diff utils folder
