+++ src/components/diff/DiffPrecompiles.tsx
@@ -1,7 +1,7 @@
 import { Address, getAddress } from 'viem';
+import { RenderDiff } from '@/components/diff/utils/RenderDiff';
 import { Copyable } from '@/components/ui/Copyable';
 import { Precompile } from '@/types';
-import { RenderDiff } from './RenderDiff';
 
 type Props = {
   base: Precompile[];
